+++ relay-server/src/processing/utils/event.rs
@@ -301,7 +301,7 @@ pub fn normalize(
             performance_issues_spans: ctx
                 .project_info
                 .has_feature(Feature::PerformanceIssuesSpans),
-            derive_trace_id: project_info.has_feature(Feature::AddDefaultTraceID),
+            force_trace_context: true,
         };
 
         metric!(timer(RelayTimers::EventProcessingNormalization), {